+++ Tests/Recurly/ShippingAddress_Test.php
@@ -21,6 +21,7 @@ class Recurly_ShippingAddressTest extends Recurly_TestCase
     $shad->zip = "94110";
     $shad->country = "US";
     $shad->company = "Recurly Inc.";
+    $shad->vat_number = "12345";
 
     return $shad;
   }
@@ -29,7 +30,7 @@ class Recurly_ShippingAddressTest extends Recurly_TestCase
     $shad = $this->mockShippingAddress();
 
     $this->assertEquals(
-      "<?xml version=\"1.0\"?>\n<shipping_address><address1>123 Dolores St.</address1><city>San Francisco</city><state>CA</state><zip>94110</zip><country>US</country><phone>555-555-5555</phone><email>verena@example.com</email><nickname>Work</nickname><first_name>Verena</first_name><last_name>Example</last_name><company>Recurly Inc.</company></shipping_address>\n",
+      "<?xml version=\"1.0\"?>\n<shipping_address><address1>123 Dolores St.</address1><city>San Francisco</city><state>CA</state><zip>94110</zip><country>US</country><phone>555-555-5555</phone><email>verena@example.com</email><nickname>Work</nickname><first_name>Verena</first_name><last_name>Example</last_name><company>Recurly Inc.</company><vat_number>12345</vat_number></shipping_address>\n",
       $shad->xml()
     );
   }
@@ -57,6 +58,20 @@ class Recurly_ShippingAddressTest extends Recurly_TestCase
         $shipping_address->update();
         $this->assertEquals($shipping_address->address1, "123 NewStreet Ave.");
       }
+    }
+  }
+
+  public function testDeleteShippingAddress() {
+    $this->client->addResponse('GET', '/accounts/abcdef1234567890/shipping_addresses', 'shipping_addresses/index-200.xml');
+    $this->client->addResponse('DELETE', 'https://api.recurly.com/v2/accounts/abcdef1234567890/shipping_addresses/1234567', 'shipping_addresses/destroy-204.xml');
+
+    $shipping_addresses = Recurly_ShippingAddressList::get('abcdef1234567890', null, $this->client);
+
+    foreach ($shipping_addresses as $shipping_address) {
+      if ($shipping_address->id == 1234567) {
+        $shipping_address->delete();
+      }
+      $this->assertInstanceOf('Recurly_ShippingAddress', $shipping_address);
     }
   }