+++ lib/recurly/shipping_address.php
@@ -9,6 +9,10 @@ class Recurly_ShippingAddress extends Recurly_Resource
     $this->_save(Recurly_Client::PUT, $this->getHref());
   }
 
+  public function delete() {
+    return Recurly_Base::_delete($this->getHref(), $this->_client);
+  }
+
   protected function getNodeName() {
     return 'shipping_address';
   }
@@ -16,7 +20,7 @@ class Recurly_ShippingAddress extends Recurly_Resource
     return array(
       'address1', 'address2', 'city', 'state',
       'zip', 'country', 'phone', 'email', 'nickname',
-      'first_name', 'last_name', 'company'
+      'first_name', 'last_name', 'company', 'vat_number'
     );
   }
   protected function populateXmlDoc(&$doc, &$node, &$obj, $nested = false) {